test: pipeline updated tests
